requirements.txt
google-cloud-storage
boto3
python-dotenv
click